+++ pantry_organizer/lib/homeLanding.dart
@@ -7,6 +7,8 @@ import 'package:flutter/src/widgets/framework.dart';
 import 'package:google_nav_bar/google_nav_bar.dart';
 import 'package:pantry_organizer/pages/foodGroup.dart';
 import 'package:pantry_organizer/pages/home.dart';
+import 'package:pantry_organizer/pages/pantry.dart';
+import 'package:pantry_organizer/pages/profile.dart';
 import 'package:pantry_organizer/pages/settings.dart';
 
 class HomeLanding extends StatefulWidget {
@@ -19,17 +21,17 @@ class HomeLanding extends StatefulWidget {
 class _HomeLandingState extends State<HomeLanding> {
   int _selectedIndex = 0;
 
-  static const List<Widget> _pages = <Widget>[
-    ProfileHome(),
-    ProfileFoodGroup(),
-    ProfileSettings()
-  ];
+  final List<Widget> _pages = <Widget>[Home(), Pantry(), ProfilePage()];
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
         appBar: AppBar(
-          title: Text('Home'),
+          title: const Text(
+            'Pantry Organizer',
+            style: TextStyle(color: Colors.white, fontSize: 20.0),
+          ),
+          centerTitle: true,
         ),
         body: Center(
           child: _pages.elementAt(_selectedIndex),
@@ -37,12 +39,13 @@ class _HomeLandingState extends State<HomeLanding> {
 
         // bottom navigation bar
         bottomNavigationBar: Container(
-          color: Colors.black,
+          color: Theme.of(context).colorScheme.primaryContainer,
           child: Padding(
             padding:
                 const EdgeInsets.symmetric(horizontal: 15.0, vertical: 20.0),
             child: GNav(
               color: Colors.white70,
+              activeColor: Theme.of(context).colorScheme.secondary,
               gap: 5,
               padding: EdgeInsets.all(16),
               tabs: const [
