+++ pantry_organizer/lib/main.dart
@@ -2,6 +2,10 @@
 import 'package:flutter/material.dart';
 import 'package:pantry_organizer/firebase_options.dart';
 import 'package:pantry_organizer/mainPage.dart';
+import 'package:pantry_organizer/pages/home.dart';
+import 'package:pantry_organizer/pages/pantry.dart';
+import 'package:pantry_organizer/pages/profile.dart';
+import 'package:pantry_organizer/pages/refrigerator.dart';
 import 'homeLanding.dart';
 import 'package:firebase_core/firebase_core.dart';
 
@@ -20,7 +24,8 @@ class MainApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      theme: ThemeData.dark().copyWith(
+      title: 'Food Pantry Organizer',
+      theme: ThemeData.light().copyWith(
         colorScheme: ColorScheme.light().copyWith(
           primary: Color.fromARGB(255, 67, 250,
               91), // easily the best line of code out of this whole damn thing 😎
@@ -31,6 +36,13 @@ class MainApp extends StatelessWidget {
         ),
       ),
       home: MainPage(),
+      //initialRoute: '/',
+      /*routes: {
+        '/': (context) => Home(),
+        'page1': (context) => const Pantry(),
+        'page2': (context) => const Refrigerator(),
+        'page3': (context) => const ProfilePage()
+      },*/
     );
   }
 }
