+++ pantry_organizer/lib/pages/home.dart
@@ -1,15 +1,73 @@
-// ignore_for_file: prefer_const_constructors, prefer_interpolation_to_compose_strings, avoid_print, prefer_const_literals_to_create_immutables, library_private_types_in_public_api
-
 import 'package:flutter/material.dart';
+import 'package:convex_bottom_bar/convex_bottom_bar.dart';
+import 'package:pantry_organizer/pages/pantry.dart';
+import 'package:pantry_organizer/pages/profile.dart';
+import 'package:pantry_organizer/pages/refrigerator.dart';
+
 class ProfileHome extends StatelessWidget {
   const ProfileHome({super.key});
 
   @override
   Widget build(BuildContext context) {
     return Center(
-      child: Text(
-        'Home Page testing'
-      ),
+      child: Text('Home Page testing'),
     );
+  }
+}
+
+class Home extends StatelessWidget {
+  static const String page0 = '/';
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+        body: Center(
+            child: Column(
+      mainAxisSize: MainAxisSize.min,
+      children: [
+        const Text(
+          'Home',
+          style: TextStyle(
+            fontWeight: FontWeight.bold,
+            fontSize: 30.0,
+          ),
+        ),
+        const SizedBox(
+          height: 30.0,
+        ),
+        ElevatedButton(
+          onPressed: () {
+            Navigator.push(
+              context,
+              MaterialPageRoute(
+                builder: (context) => const Pantry(),
+              ),
+            );
+          },
+          child: const Text('Pantry'),
+        ),
+        ElevatedButton(
+          onPressed: () {
+            Navigator.push(
+              context,
+              MaterialPageRoute(
+                builder: (context) => const Refrigerator(),
+              ),
+            );
+          },
+          child: const Text('Refrigerator'),
+        ),
+        ElevatedButton(
+          onPressed: () {
+            Navigator.push(
+              context,
+              MaterialPageRoute(
+                builder: (context) => const ProfilePage(),
+              ),
+            );
+          },
+          child: const Text('Profile'),
+        ),
+      ],
+    )));
   }
 }
