+++ pantry_organizer/lib/pages/subpages/groupPage.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart';
+
+class GroupPage extends StatelessWidget {
+  const GroupPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    // TODO: implement build
+    throw UnimplementedError();
+  }
+}
