+++ pantry_organizer/lib/utils/UserPreferences.dart
@@ -1,0 +1,11 @@
+import 'package:firebase_auth/firebase_auth.dart';
+
+class UserPreferences {
+  static const myUser = User(
+    imagePath: "",
+    name: 'Jacob >:(',
+    email: '',
+    about: "",
+    isDarkMode: true,
+  );
+}
