+++ pantry_organizer/lib/widget/AppBar.dart
@@ -1,0 +1,28 @@
+import 'package:convex_bottom_bar/convex_bottom_bar.dart';
+import 'package:flutter/material.dart';
+
+class CustomAppBar extends AppBar {
+  CustomAppBar(BuildContext context, {super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: Text('')),
+      body: Center(
+          child: TextButton(
+        child: Text(' '),
+        onPressed: () => Navigator.of(context).pushNamed('/bar'),
+      )),
+      bottomNavigationBar: ConvexAppBar(
+        style: TabStyle.react,
+        items: [
+          TabItem(icon: Icons.home, title: 'Home'),
+          TabItem(icon: Icons.fastfood_sharp, title: 'Pantry'),
+          TabItem(icon: Icons.person, title: 'Profile')
+        ],
+        initialActiveIndex: 1,
+        onTap: (int i) => print('click index=$i'),
+      ),
+    );
+  }
+}
