+++ pantry_organizer/pubspec.yaml
@@ -39,6 +39,7 @@ dependencies:
   cloud_firestore: 
   flutterfire_cli:
   google_nav_bar: 
+  convex_bottom_bar: ^3.1.0+1
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
