--- utils/UserPreferences.dart
@@ -1,11 +1,0 @@
-import 'package:user_profile_example/model/user.dart';
-
-class UserPreferences {
-  static const myUser = User(
-    imagePath:
-    name: 'Jacob >:(',
-    email: '',
-    about:
-    isDarkMode: ,
-  );
-}
